+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fix for calls to `evaluate` that return JSON: only set `serialization_options` when `return_by_value=False` @thromer
+
 ### Added
 
 ### Changed